Don't pass parseParam to wrapTracking.
The parseParam function is just a helper that returns a boolean value, rather
than a true parsing function that returns an AST node. When wrapped by
wrapTracking, which assumes that the result of parseFunction is an AST node,
parseParam becomes unusable in browsers that restrict property accesses on
boolean values.