Adopt semver-compliant versioning that preserves both local and upstream versions
This fork of esprima, for the foreseeable future, only adds features to its
upstream counterpart. As such, we would like it to be possible to track both
changes to the upstream version it is tracking as well as local changes (in a
semver compliant way).

We've decided that the way we will do this is to simply make all semver version
parts equal to (fb-harmony version) * 1000 * (upstream version). If upstream
esprima ever gets to version 999, I'll admit defeat...but hopefully that's an
unlikely possibility.